# [WIP] Add include_for_find support via 'expand'