@Fryguy 2020-07-29
Nice!  Do you have some performance comparisons?