@jrafanie 2020-07-29
> Nice! Do you have some performance comparisons?

@NickLaMuro do we have to opt-in or something?  I don't see improvements with applying this change on this request:

Before:
`bundle exec miqperf benchmark -ac 5  "/api/vms?expand=resources&attributes=num_cpu,name"`

```
/api/vms
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 2054 |    1116 |     1011.9 | 1665 |
| 2299 |    1116 |     1092.3 | 1665 |
| 2031 |    1116 |      987.9 | 1665 |
| 1952 |    1116 |      939.9 | 1665 |
| 1934 |    1116 |      943.0 | 1665 |
```

After:
(add `expand=hardware`)

`bundle exec miqperf benchmark -ac 5  "/api/vms?expand=resources,hardware&attributes=num_cpu,name"`

```
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 3798 |    1118 |     1092.4 | 3112 |
| 2091 |    1116 |     1019.4 | 1665 |
| 2332 |    1116 |     1090.9 | 1665 |
| 2054 |    1116 |      972.1 | 1665 |
| 1973 |    1116 |      933.7 | 1665 |
| 2082 |    1116 |      990.3 | 1665 |
| 2092 |    1116 |      973.7 | 1665 |
| 2089 |    1116 |      978.1 | 1665 |
| 2320 |    1116 |     1076.9 | 1665 |
| 2098 |    1116 |      984.0 | 1665 |
```