@abellotti 2020-07-29
@jrafanie you won't see an improvement with that request (attributes=...) as this enhancement is for expanding subcollections.  i.e. expand=...  see method used https://github.com/ManageIQ/manageiq-api/pull/877/files#diff-8caecf6616f6fd1e2602842ad7dd51f0R108.  Once this is expanded to also include virtual attributes and associations (via attributes), then you'd see the improvement with your query.