@NickLaMuro 2020-07-29
@abellotti no, I don't think that is correct.

The N+1 was happening because `num_cpu` is a `virtual_delegate` on `hardware`, and so it has to fetch it individually for each `VM` record when it gets here:

https://github.com/ManageIQ/manageiq-api/blob/master/app/controllers/api/base_controller/renderer.rb#L214

However, I think the issue is that @jrafanie is seeing I was probably testing locally with this patch also applied:

https://github.com/ManageIQ/manageiq-api/pull/874

So I think they were working together to get a better result overall, and without it, the N+1 causing a refetch of the VMs obliterates any `.includes` that might have been included early on in the `.collection_search` method.

This report output is a single run where I was modifying a few things in place:

```
$ bundle exec miqperf report --last
/api/vms
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 2453 |     565 |      193.8 | 2594 |
| 2852 |     566 |      208.6 | 2595 |
| 2930 |     566 |      240.7 | 2595 |
| 2290 |      13 |         79 | 2042 |
```

And the last one, I tweaked the `.reject` from being a `!=` (incorrect) to `==` (correct), and that caused the queries to drop down to what they should be.

* * *

I will mark this PR as `[WIP]` while I double check my theory, and also to wait for #874 to be merge so I can add tests without conflicts