@abellotti 2020-07-29
```GET /api/vms/:id?expand=hardware``` does not return hardware. (hardware is not a subcollection)
```GET /api/vms/:id?attributes=hardware``` will return hardware


If the intent is to know what to prefetch for a virtual attributes specified in attributes, looking at expand does not seem right as callers will not be specifying those via expand (only for will defined subcollections).  Is there a way to find what to preload for a virtual attribute specified via attributes ?