@NickLaMuro 2020-07-29
@abellotti Thanks, I figured you would have had some issue with how I was designing this 😄 

> If the intent is to know what to prefetch for a virtual attributes specified in attributes

Not really, or at least it wasn't my intent, though it could be.

The idea was more that this was overloading `expand` to allow for an opt-in for an `.includes`, without resorting to adding a new URL param (maybe called `"includes="` or something similar).  However, I could be convinced that it makes more sense to do that to avoid muddying the use case of `expand=` like I am doing here.