@NickLaMuro 2020-07-29
> Is there a way to find what to preload for a virtual attribute specified via attributes ?

I can investigate what this would take more today, though, it is significantly harder to do with `virtual_attributes` than just a simple `hardware.cpu_sockets`, where the table name is right there when we need it.

This was the explicit approach, where a user has to opt in to do it, where the later would be more complex, but would be automatically applied.