@NickLaMuro 2020-07-29
> Nice! Do you have some performance comparisons?

@Fryguy Yeah, sorry, I mentioned some when responding to @jrafanie here:

https://github.com/ManageIQ/manageiq-api/pull/877#issuecomment-665769849

But I will throw some more formal ones together and put them in the PR description