@NickLaMuro 2020-07-29
As LJ showed, without my patch from https://github.com/ManageIQ/manageiq-api/pull/874 in place, I got the following for performance numbers:

```console
$ bundle exec miqperf benchmark -ac 5 "/api/vms?expand=resources,hardware&attributes=num_cpu,name"
D, [2020-07-29T15:16:48.277412 #13338] DEBUG -- : --> logging in...
D, [2020-07-29T15:16:54.638950 #13338] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:16:59.451455 #13338] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:17:02.135731 #13338] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:17:04.765458 #13338] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:17:07.750053 #13338] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
$ bundle exec miqperf report --last
/api/vms
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 4241 |    1118 |      527.5 | 3147 |
| 2192 |    1116 |      460.4 | 1665 |
| 2132 |    1116 |      512.0 | 1665 |
| 2463 |    1116 |      469.7 | 1665 |
| 2592 |    1116 |      549.0 | 1665 |
```

However, after applying the patch manually to that branch in the API:

```console
$ git apply <(curl -L https://github.com/ManageIQ/manageiq-api/pull/874.patch)
```

I got the numbers I expected:


```console
$ bundle exec miqperf benchmark -ac 5 "/api/vms?expand=resources,hardware&attributes=num_cpu,name"
D, [2020-07-29T15:18:04.040054 #13731] DEBUG -- : --> logging in...
D, [2020-07-29T15:18:09.667021 #13731] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:18:11.642624 #13731] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:18:12.000060 #13731] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:18:12.357015 #13731] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
D, [2020-07-29T15:18:12.777287 #13731] DEBUG -- : --> making GET request: /api/vms?expand=resources,hardware&attributes=num_cpu,name
$ bundle exec miqperf report --last
/api/vms
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 1738 |      12 |         35 | 2041 |
|  260 |      10 |        9.8 |  559 |
|  262 |      10 |        9.2 |  559 |
|  257 |      10 |        9.6 |  559 |
|  259 |      10 |       10.0 |  559 |
```