@jrafanie 2020-07-31
I've confirmed the patches on jansa improve the results as @NickLaMuro reported...

```
$ git apply <(curl -L https://github.com/ManageIQ/manageiq-api/pull/874.patch)
$ git apply <(curl -L https://github.com/ManageIQ/manageiq-api/pull/877.patch)
```